+++ assets/style/styles.css
@@ -1,0 +1,281 @@
+@import url("https://fonts.googleapis.com/css2?family=IBM+Plex+Sans+KR&display=swap");
+
+/* 전체 폰트 설정 */
+* {
+    font-family: "IBM Plex Sans KR", sans-serif;
+    font-weight: 400;
+    font-style: normal;
+}
+
+/* 배경색 설정 */
+body {
+    background-color: #5385C1;
+    text-align: center;
+}
+
+/* 헤더 섹션 스타일 */
+header {
+    position: absolute;
+    width: 100%;
+    top: 0;
+    left: 0;
+    background: rgba(0, 0, 0, 0.0);
+    /* 투명한 배경 */
+    z-index: 100;
+    padding-top: 5vh, 0, 0, 0, ;
+    margin: 5vh, 0, 0, 0;
+}
+
+header .nav-link {
+    color: white !important;
+}
+
+/* 메인 배경 설정 */
+.main {
+    color: white;
+    background-image: url("image/banner.jpg");
+    background-position: top;
+    background-size: cover;
+    background-attachment: fixed;
+    background-repeat: no-repeat;
+    min-height: 40vh;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+
+.main .container-fluid {
+    background-color: rgba(0, 0, 0, 0.0);
+    border-radius: 10px;
+    padding: 20px;
+    align-items: center;
+}
+
+/* 멤버 추가 입력 폼 스타일 */
+.mypostingbox {
+    width: 500px;
+    margin: 20px auto;
+    padding: 20px;
+    border: 1px solid #ccc;
+    border-radius: 10px;
+    display: none;
+    background-color: transparent;
+}
+
+.form-floating>input {
+    background-color: transparent;
+    color: white;
+}
+
+.form-floating>label {
+    color: white;
+}
+
+.form-control:focus {
+    color: white;
+    background-color: transparent;
+}
+
+.form-control:focus {
+    background-color: transparent;
+}
+
+.mybtn {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-end;
+}
+
+.mybtn>button {
+    margin-right: 5px;
+    width: 100%;
+}
+
+/* 카드 스타일 */
+.mycards {
+    width: 90%;
+    margin: 20px auto;
+}
+
+/* 카드 래퍼 스타일 */
+.mycards-wrapper {
+    padding: 2rem 5rem 5rem 5rem;
+    background-color: rgba(255, 255, 255, 0.0);
+    border-radius: 10px;
+}
+
+.card-section-text {
+    padding: 1rem 10rem 1rem 10rem;
+    color: white;
+}
+
+/* 카드 스타일 수정 */
+.card {
+    width: 80%;
+    height: 80%;
+    border-radius: 10px;
+    border-color: rgba(190, 190, 190, 0.5);
+    background-color: transparent;
+    /* 카드 배경색 */
+    transition: transform 0.2s ease-in-out;
+    margin: 1rem;
+    /* 카드 간 간격 */
+    padding: 2rem;
+    text-align: center;
+}
+
+.card:hover {
+    transform: translateY(-10px);
+}
+
+.card-img-wrapper {
+    width: 100%;
+    padding-top: 100%;
+    position: relative;
+    border-radius: 50%;
+    overflow: hidden;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin-bottom: 1rem;
+    /* Center the image and add margin at the bottom */
+}
+
+.card-img-top {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+
+.card-body {
+    text-align: center;
+    padding: 1rem;
+}
+
+.card-title {
+    font-size: 2rem;
+    margin-bottom: 0.5rem;
+    color: white;
+}
+
+.card-text {
+    font-size: 1.25rem;
+    color: #777;
+    color: white;
+}
+
+/* 모달 스타일 */
+.modal-content {
+    background-color: #8fb0da;
+    color: black;
+    border-radius: 10px;
+}
+
+.modal-header,
+.modal-footer {
+    border: none;
+}
+
+.modal-body .text-left {
+    text-align: left;
+}
+
+.modal-footer .link-container {
+    text-align: left;
+}
+
+.modal-footer .link-list {
+    padding-left: 0;
+    list-style: none;
+}
+
+.modal-footer .col-md-8 {
+    display: flex;
+    align-items: center;
+}
+
+.modal-footer .col-md-4 {
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+}
+
+.modal-footer .btn {
+    margin-right: 5px;
+}
+
+.link-container {
+    text-align: left;
+    padding: 0 15px;
+}
+
+.link-list {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+    display: flex;
+    gap: 15px;
+}
+
+.link-list li {
+    display: inline;
+}
+
+.link-list a {
+    color: #007bff;
+    text-decoration: none;
+}
+
+.link-list a:hover {
+    text-decoration: underline;
+}
+
+.btn-primary {
+    background-color: #007bff;
+    border-color: #007bff;
+}
+
+.btn-primary:hover {
+    background-color: #0056b3;
+    border-color: #0056b3;
+}
+
+.btn-outline-light {
+    color: #f8f9fa;
+    border-color: #f8f9fa;
+}
+
+.btn-outline-light:hover {
+    color: #212529;
+    background-color: #f8f9fa;
+    border-color: #f8f9fa;
+}
+
+.btn-danger {
+    background-color: #dc3545;
+    border-color: #dc3545;
+}
+
+.btn-danger:hover {
+    background-color: #c82333;
+    border-color: #bd2130;
+}
+
+.btn-text-only {
+    background-color: transparent;
+    border: none;
+    color: #f8f9fa;
+    padding: 0;
+    font-size: 1rem;
+}
+
+.nav-link-button {
+    padding: 0.5rem 1rem;
+    color: #f8f9fa;
+    text-decoration: none;
+    cursor: pointer;
+}
