--- styles.css
@@ -1,121 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=IBM+Plex+Sans+KR&display=swap");
-
-/* 전체 폰트 설정 */
-* {
-    font-family: "IBM Plex Sans KR", sans-serif;
-    font-weight: 400;
-    font-style: normal;
-}
-
-/* 배경색 설정 */
-body {
-    background-color: #f5f5f5;
-}
-
-/* 메인 배경 이미지 설정 */
-.main {
-    color: white;
-    background-image: url("https://images.unsplash.com/photo-1496389394391-c23f6d8a9954?ixlib=rb-1.2.1&auto=format&fit=crop&w=1650&q=80");
-    background-position: center;
-    background-size: cover;
-}
-
-.main .container-fluid {
-    background-color: rgba(0, 0, 0, 0.6);
-    border-radius: 10px;
-}
-
-/* 멤버 추가 입력 폼 스타일 */
-.mypostingbox {
-    width: 500px;
-    margin: 20px auto;
-    padding: 20px;
-    border: 1px solid #ccc;
-    border-radius: 10px;
-    display: none;
-    background-color: #fff;
-}
-
-.form-floating>input {
-    background-color: transparent;
-    color: black;
-}
-
-.form-floating>label {
-    color: #555;
-}
-
-.mybtn {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-}
-
-.mybtn>button {
-    margin-right: 5px;
-    width: 100%;
-}
-
-/* 카드 스타일 */
-.mycards {
-    width: 90%;
-    margin: 20px auto;
-}
-
-.card {
-    border-radius: 15px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-.card-img-top {
-    width: 100%;
-    height: auto;
-    max-height: 700px;
-    object-fit: cover;
-    border-top-left-radius: 15px;
-    border-top-right-radius: 15px;
-}
-
-/* 모달 스타일 */
-.modal-content {
-    background-color: #fff;
-    color: black;
-    border-radius: 10px;
-}
-
-.modal-header,
-.modal-body,
-.modal-footer {
-    border: none;
-}
-
-.btn-primary {
-    background-color: #007bff;
-    border-color: #007bff;
-}
-
-.btn-primary:hover {
-    background-color: #0056b3;
-    border-color: #0056b3;
-}
-
-.btn-outline-light {
-    color: #f8f9fa;
-    border-color: #f8f9fa;
-}
-
-.btn-outline-light:hover {
-    color: #212529;
-    background-color: #f8f9fa;
-    border-color: #f8f9fa;
-}
-
-.btn-danger {
-    background-color: #dc3545;
-    border-color: #dc3545;
-}
-
-.btn-danger:hover {
-    background-color: #c82333;
-    border-color: #bd2130;
-}
